requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
spacy
scikit-learn
jinja2